move files to root
