+++ binary-to-bytea-stream.js
