+++ bytea-to-binary-stream.js
