+++ bytea-to-binary.js
